Introduce an enabled flag for S/MIME
This allows setups where there is no S/MIME. In some scenarios using
just GPG is fine and S/MIME might even be discouraged. Previously we had
to provide a dummy S/MIME key just to make remail happy. With this new
flag we do not need that key if we do not intend to use it.